feat: - flag for disabling pokevial - add a use every time you get a badge/when you become champ/beat steven - time-based mugshot colors (will be adding more) fix: - surf was acting weird @ route 115

+++ include/data.h
@@ -271,10 +271,11 @@ static inline const u8 GetTrainerPartySizeFromId(u16 trainerId)
 
 static inline const bool32 DoesTrainerHaveMugshot(u16 trainerId)
 {
+    
     return gTrainers[GetCurrentDifficultyLevel()][SanitizeTrainerId(trainerId)].mugshotColor;
 }
 
-static inline const u8 GetTrainerMugshotColorFromId(u16 trainerId)
+static inline const u32 GetTrainerMugshotColorFromId(u16 trainerId)
 {
     return gTrainers[GetCurrentDifficultyLevel()][SanitizeTrainerId(trainerId)].mugshotColor;
 }
